update method wip
